+++ tests/help/test_name.py
@@ -13,8 +13,10 @@ from tests.utils import parse, strip_trailing_whitespace
 def test_argument_name(capsys):
     @dataclass
     class Args:
-        name: Annotated[str, cappa.Arg(value_name="string-name", help="more")]
-        short: Annotated[str, cappa.Arg(short=True, value_name="optional-string")]
+        name: Annotated[str, cappa.Arg(value_name="sname", help="more")]
+        short: Annotated[str, cappa.Arg(short=True, value_name="ostr")]
+        unbounded: Annotated[list[list[str]], cappa.Arg(short=True, value_name="UNB")]
+        unbounded_pos: Annotated[list[str], cappa.Arg(value_name="upos", help="lots")]
 
     with pytest.raises(cappa.HelpExit) as e:
         parse(Args, "--help")
@@ -25,13 +27,15 @@ def test_argument_name(capsys):
 
     assert out == textwrap.dedent(
         """\
-        Usage: args -s OPTIONAL-STRING STRING-NAME [-h] [--completion COMPLETION]
+        Usage: args -s OSTR -u UNB ... SNAME UPOS ... [-h] [--completion COMPLETION]
 
           Options
-            -s OPTIONAL-STRING
+            -s OSTR
+            -u UNB ...
 
           Arguments
-            STRING-NAME                more
+            SNAME                      more
+            UPOS ...                   lots
 
           Help
             [-h, --help]               Show this message and exit.